move player store back to board
