--- cambio-client/src/store/player/actions.js
@@ -1,18 +1,0 @@
-import { action } from 'typesafe-actions'
-import { mockFetch } from '../../mocks/fetch'
-
-import {
-    START_GAME,
-    STARTING_GAME
-} from './types'
-
-export const startGame = () => {
-    return dispatch => {
-        dispatch(action(STARTING_GAME))
-
-        mockFetch("/deal")
-            .then(players => {
-                dispatch(action(START_GAME, players))
-            })
-    }
-}
