--- cambio-client/src/store/player/selectors.js
@@ -1,1 +1,0 @@
-export const getPlayers = ({ players }) => players
