--- cambio-client/src/store/player/types.js
@@ -1,2 +1,0 @@
-export const START_GAME = "@@board/START_GAME"
-export const STARTING_GAME = "@@board/STARTING_GAME"
